json 리턴하는 함수는 xxApiController로 분리, chart 오류 해결

+++ src/main/java/com/example/springsecuritytest/batch/SimpleJobConfiguration.java
@@ -100,7 +100,7 @@ public class SimpleJobConfiguration {
             // batchresult로 만드는 작업
             return BatchResult.builder()
                     .name(objToString.get(0))
-                    .bbsCount(Long.parseLong(objToString.get(1)))
+                    .bbsCount(Integer.parseInt(objToString.get(1)))
                     .staticsDate(beforeDate)
                     .build();
         };
